// SIC/SIC Admin/src/types/component.d.ts
export type CheckBoxTreeProps = {
	nodes: {
		value: string;
		label: string;
		id: string;
		children: {
			value: string;
			label: string;
		}[];
	}[];
	checkedChild: string[];
	onChangeCheckedChildHandler: (value) => void;
	columns?: number;
};